Add files via upload
when you run the code, you will need to sign up/make an account then sign in. after that click on the camera icon in the home screen, that will take you to basically a little concept of the cmaera functionality, but below that there is a button where the user will upload a picture. thats where you guys can add ur code. in the actual code it should be from 478 to 524. look for cameraView