Add files via upload
when you run the code, you will need to sign up/make an account then sign in. after that click on the camera icon in the home screen, that will take you to basically a little concept of the cmaera functionality, but below that there is a button where the user will upload a picture. thats where you guys can add ur code. in the actual code it should be from 478 to 524. look for cameraView.the clock in the middle will take you to a new page where you will see like example of the users shopping history and the gear will take you to the user profile setting and when clicking the buttons they are all self explanitory in what they do and you will be able to clearly see what happens